client state progress, initial dashboard ssr and clean csr handover

+++ src/lib/server/session.js
@@ -1,6 +1,6 @@
-import db from "./db.js";
-import { token } from "./util/token.js";
-import { unix } from "./util/unix.js";
+import db from './db.js';
+import { token } from './util/token.js';
+import { unix } from './util/unix.js';
 
 const SESSION_TIMEOUT_ABSOLUTE = 60 * 60 * 24 * 14; // two weeks
 const SESSION_TIMEOUT_INACTIVITY = 60 * 60 * 24; // one day
@@ -15,10 +15,12 @@ export const SESSION_MAX_AGE = SESSION_TIMEOUT_ABSOLUTE;
 export function createSession(user_id) {
 	const session_id = token();
 	const unix_now = unix();
-	db.exec(
-		"INSERT INTO sessions (id, created_at, updated_at, user_id) VALUES (?, ?, ?, ?)",
-		[session_id, unix_now, unix_now, user_id],
-	);
+	db.exec('INSERT INTO sessions (id, created_at, updated_at, user_id) VALUES (?, ?, ?, ?)', [
+		session_id,
+		unix_now,
+		unix_now,
+		user_id
+	]);
 	return session_id;
 }
 
@@ -37,7 +39,7 @@ export function getSessionUserInfo(session_id) {
 		FROM sessions
 		LEFT JOIN users ON sessions.user_id = users.id
 		WHERE sessions.id = ? AND sessions.created_at > ? AND sessions.updated_at > ?`,
-		[firstLoginCutoff, session_id, absoluteTimeout, inactivityTimeout],
+		[firstLoginCutoff, session_id, absoluteTimeout, inactivityTimeout]
 	);
 	if (!user) {
 		return null;
@@ -51,10 +53,7 @@ export function getSessionUserInfo(session_id) {
  */
 export function updateSession(session_id) {
 	const unix_now = unix();
-	db.exec("UPDATE sessions SET updated_at = ? WHERE id = ?", [
-		unix_now,
-		session_id,
-	]);
+	db.exec('UPDATE sessions SET updated_at = ? WHERE id = ?', [unix_now, session_id]);
 }
 
 /**
@@ -62,7 +61,7 @@ export function updateSession(session_id) {
  * @param {string} session_id
  */
 export function invalidateSession(session_id) {
-	db.exec("DELETE FROM sessions WHERE id = ?", [session_id]);
+	db.exec('DELETE FROM sessions WHERE id = ?', [session_id]);
 }
 
 /**
@@ -71,8 +70,5 @@ export function invalidateSession(session_id) {
  * @param {string} session_id
  */
 export function invalidateOtherSessions(user_id, session_id) {
-	db.exec("DELETE FROM sessions WHERE user_id = ? AND id != ?", [
-		user_id,
-		session_id,
-	]);
+	db.exec('DELETE FROM sessions WHERE user_id = ? AND id != ?', [user_id, session_id]);
 }
